Rename module from roadshow_diffusion to diffusion_model
