+++ diffusion_model.py
