+++ pyproject.toml
@@ -6,7 +6,7 @@ requires = [
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "roadshow-diffusion"
+name = "diffusion-model"
 version = "0.5"
 dependencies = [
     "matplotlib",
@@ -15,4 +15,4 @@ dependencies = [
 ]
 
 [tool.setuptools]
-py-modules = ["roadshow_diffusion"]
+py-modules = ["diffusion_model"]
